requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
python-pptx
convertapi
supabase
python-dotenv
livekit
livekit-agents
livekit-plugins-google
livekit-plugins-anam